Example update, Add more user functionality

+++ tests/integration/User.tests.ps1
@@ -1,13 +1,24 @@
-. (Join-Path $PSScriptRoot '../setup.ps1')
+BeforeAll {
+    . (Join-Path $PSScriptRoot '../setup.ps1')
+}
 
 # Pester 5 need to pass in TestCases object to pass share
 $PSDefaultParameterValues = @{
     "It:TestCases" = @{
                         email = 'testuser@mailinator.com'
                         email2 = 'testuser2@mailinator.com'
+                        email3 = 'testuser3@mailinator.com'
+                        email4 = 'testuser4@mailinator.com'
+                        email5 = 'testuser5@mailinator.com'
+                        email6 = 'testuser6@mailinator.com'
                         vars = @{
                             user = $null
                             user2 = $null
+                            user3 = $null
+                            user4= $null
+                            user5 = $null
+                            user6 = $null
+                            group = $null
                             userCount = 0
                         }
                     }
@@ -17,6 +28,14 @@ Describe "Cleanup" {
     It "Remove test user" {
         (Get-OktaUser -q $email) | Remove-OktaUser -Confirm:$false
 	    (Get-OktaUser -q $email2) | Remove-OktaUser -Confirm:$false
+	    (Get-OktaUser -q $email3) | Remove-OktaUser -Confirm:$false
+	    (Get-OktaUser -q $email4) | Remove-OktaUser -Confirm:$false
+	    (Get-OktaUser -q $email5) | Remove-OktaUser -Confirm:$false
+	    (Get-OktaUser -q $email6) | Remove-OktaUser -Confirm:$false
+        $vars.group = Get-OktaGroup -Query "userTestGroup"
+        if (!$vars.group) {
+            $vars.group = New-OktaGroup -Name "userTestGroup"
+        }
     }
 }
 
@@ -26,10 +45,33 @@ Describe "User" {
         $vars.user | Should -Not -Be $null
         $vars.user.Status | Should -Be 'STAGED'
     }
+    It "Adds a user with recovery question" {
+        $vars.user3 = New-OktaUser -FirstName test-user -LastName test-user -Email $email3 -RecoveryQuestion Why? -RecoveryAnswer "Answer is 42"
+        $vars.user3 | Should -Not -Be $null
+        $vars.user3.Status | Should -Be 'STAGED'
+        $vars.user3.credentials.recovery_question.question | Should -Be 'Why?'
+    }
+    It "Adds a user to login next" {
+        $vars.user4 = New-OktaUser -FirstName test-user -LastName test-user -Email $email4 -NextLogin -Activate -Pw "12SHiwS9876$%#"
+        $vars.user4 | Should -Not -Be $null
+        $vars.user4.Status | Should -Be 'PASSWORD_EXPIRED'
+    }
+    It "Adds a user to a group" {
+        $vars.user5 = New-OktaUser -FirstName test-user -LastName test-user -Email $email5 -GroupIds @($vars.group.id)
+        $vars.user5 | Should -Not -Be $null
+        $vars.user5.Status | Should -Be 'STAGED'
+        $groups = Get-OktaUserGroup -UserId $vars.user5.id
+        $groups.id -contains $vars.group.id | Should -Be $true
+    }
     It "Adds AuthProviderUser" {
         $vars.user2 = New-OktaAuthProviderUser -FirstName "fn" -LastName "ln" -Email $email2 -ProviderType FEDERATION -ProviderName FEDERATION -Activate
         $vars.user2 | Should -Not -Be $null
         $vars.user2.status | Should -Be 'ACTIVE'
+    }
+    It "Adds AuthProviderUser Activated" {
+        $vars.user6 = New-OktaAuthProviderUser -FirstName "fn" -LastName "ln" -Email $email6 -ProviderType SOCIAL -Activate
+        $vars.user6 | Should -Not -Be $null
+        $vars.user6.status | Should -Be 'PROVISIONED'
     }
     It "Gets Users" {
         $result = @(Get-OktaUser)
@@ -89,8 +131,13 @@ Describe "User" {
 
 Describe "Cleanup" {
     It "Remove test user" {
-        Remove-OktaUser -UserId $vars.user.id -Confirm:$false
-        Remove-OktaUser -UserId $vars.user2.id -Confirm:$false
+        if ($vars.user) { Remove-OktaUser -UserId $vars.user.id -Confirm:$false }
+        if ($vars.user2) { Remove-OktaUser -UserId $vars.user2.id -Confirm:$false }
+        if ($vars.user3) { Remove-OktaUser -UserId $vars.user3.id -Confirm:$false }
+        if ($vars.user4) { Remove-OktaUser -UserId $vars.user4.id -Confirm:$false }
+        if ($vars.user5) { Remove-OktaUser -UserId $vars.user5.id -Confirm:$false }
+        if ($vars.user6) { Remove-OktaUser -UserId $vars.user6.id -Confirm:$false }
+        if ($vars.group) { Remove-OktaGroup -GroupId $vars.group.id -Confirm:$false }
     }
 }
